criando migrates (V2)

+++ database/migrations/2024_06_19_224706_create_unidades_table.php
@@ -21,12 +21,12 @@ class CreateUnidadesTable extends Migration
         });
         //adicionar o relacionamento com a tabela produtos
         Schema::table('produtos', function (Blueprint $table){
-             $table->unsignedInteger('unidade_id');
+             $table->unsignedBigInteger('unidade_id');
              $table->foreign('unidade_id')->references('id')->on('unidades');   
         });
         //adicionar o relacionamento com a tabela produto_detalhes
         Schema::table('produto_detalhes', function (Blueprint $table){
-            $table->unsignedInteger('unidade_id');
+            $table->unsignedBigInteger('unidade_id');
             $table->foreign('unidade_id')->references('id')->on('unidades');   
         });
